@@ -6,7 +6,7 @@ module Comunit
   class Engine < ::Rails::Engine
     config.generators do |g|
       g.test_framework :rspec
-      g.fixture_replacement :factory_bot, :dir => 'spec/factories'
+      g.fixture_replacement :factory_bot, dir: 'spec/factories'
     end
 
     initializer 'sample_engine.factories', after: 'factory_bot.set_factory_paths' do
